requirements.txt
pytesseract == 0.3.13
pymupdf == 1.25.2
python-docx == 1.1.2
pdfplumber == 0.11.5
pillow == 11.1.0
transformers == 4.48.1
torch == 2.5.1
torchvision == 0.20.1
git+https://github.com/openai/CLIP.git
openai == 1.60.1
python-dotenv == 1.0.1